stats + fixes

+++ RecruitmentSystem/RecruitmentSystem.Domain/Dtos/Decision/StepEvaluation.cs
@@ -3,7 +3,7 @@ namespace RecruitmentSystem.Domain.Dtos.Decision;
 public class StepEvaluation
 {
     public string StepName { get; set; }
-    public int AiScoreForCandidateInStep { get; set; }
-    public int CompanyScoreForCandidateInStep { get; set; }
+    public double AiScoreForCandidateInStep { get; set; }
+    public double CompanyScoreForCandidateInStep { get; set; }
     public string CompanysReviewOfCandidateInStep { get; set; }
 }
